requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# dishq/urls.py: 17
# dishq/wsgi.py: 12
# image_processing/admin.py: 1
# image_processing/apps.py: 1
# image_processing/migrations/0001_initial.py: 3
# image_processing/models.py: 1,2
# image_processing/tests.py: 1,2,3
# image_processing/urls.py: 1
# image_processing/views.py: 1,2,3,4
# manage.py: 8
Django == 2.0.3

# image_processing/models.py: 5
Pillow == 4.2.1

# image_processing/tests.py: 5,6
# image_processing/views.py: 6,7,8
djangorestframework == 3.7.7
